ref: Переместил определение клавиатуры в отдельный файл
